[#156221725] Refactored test directory structure
